+++ client/python/openlineage/client/transport/msk_iam.py
@@ -15,7 +15,7 @@ log = logging.getLogger(__name__)
 
 def _detect_running_region() -> None | str:
     """Dynamically determine the region."""
-    import boto3  # type: ignore[import]
+    import boto3  # type: ignore[import-untyped]
 
     checks = [
         # check if set through ENV vars
@@ -44,7 +44,7 @@ class MSKIAMConfig(KafkaConfig):
 
 
 def _oauth_cb(config: MSKIAMConfig, *_: Any) -> tuple[str, float]:
-    from aws_msk_iam_sasl_signer import MSKAuthTokenProvider  # type: ignore[import]
+    from aws_msk_iam_sasl_signer import MSKAuthTokenProvider  # type: ignore[import-untyped]
 
     region = config.region
     if config.aws_profile: