+++ integration/spark/app/src/test/java/io/openlineage/spark/agent/lifecycle/plan/CustomEnvironmentVariablesCaptureTest.java
@@ -87,7 +87,7 @@ class CustomEnvironmentVariablesCaptureTest {
                     e.getEventType().equals(EventType.START)
                         && e.getJob().getFacets() != null
                         // Exclude SparkApplication start event because env variable was set after
-                        && e.getJob().getFacets().getJobType() != null
+                        && e.getJob().getFacets().getJobType().getJobType() != "APPLICATION"
                         && e.getRun().getFacets() != null
                         && e.getRun()
                                 .getFacets()