+++ integration/spark/shared/src/main/java/io/openlineage/spark/agent/lifecycle/OpenLineageRunEventBuilder.java
@@ -14,14 +14,11 @@ import io.openlineage.client.OpenLineage.DatasetFacets;
 import io.openlineage.client.OpenLineage.InputDataset;
 import io.openlineage.client.OpenLineage.InputDatasetFacet;
 import io.openlineage.client.OpenLineage.InputDatasetInputFacets;
-import io.openlineage.client.OpenLineage.JobBuilder;
 import io.openlineage.client.OpenLineage.JobFacet;
 import io.openlineage.client.OpenLineage.OutputDataset;
 import io.openlineage.client.OpenLineage.OutputDatasetFacet;
 import io.openlineage.client.OpenLineage.OutputDatasetOutputFacets;
-import io.openlineage.client.OpenLineage.ParentRunFacet;
 import io.openlineage.client.OpenLineage.RunEvent;
-import io.openlineage.client.OpenLineage.RunEventBuilder;
 import io.openlineage.client.OpenLineage.RunFacet;
 import io.openlineage.client.OpenLineage.RunFacets;
 import io.openlineage.client.OpenLineage.RunFacetsBuilder;
@@ -35,8 +32,6 @@ import io.openlineage.spark.api.CustomFacetBuilder;
 import io.openlineage.spark.api.OpenLineageContext;
 import io.openlineage.spark.api.OpenLineageEventHandlerFactory;
 import io.openlineage.spark.api.QueryPlanVisitor;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
@@ -51,17 +46,11 @@ import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.spark.rdd.RDD;
 import org.apache.spark.scheduler.ActiveJob;
-import org.apache.spark.scheduler.SparkListenerApplicationEnd;
-import org.apache.spark.scheduler.SparkListenerApplicationStart;
 import org.apache.spark.scheduler.SparkListenerEvent;
-import org.apache.spark.scheduler.SparkListenerJobEnd;
 import org.apache.spark.scheduler.SparkListenerJobStart;
 import org.apache.spark.scheduler.SparkListenerStageCompleted;
-import org.apache.spark.scheduler.SparkListenerStageSubmitted;
 import org.apache.spark.scheduler.Stage;
 import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan;
-import org.apache.spark.sql.execution.ui.SparkListenerSQLExecutionEnd;
-import org.apache.spark.sql.execution.ui.SparkListenerSQLExecutionStart;
 import scala.Function1;
 import scala.PartialFunction;
 
@@ -194,181 +183,15 @@ class OpenLineageRunEventBuilder {
                 }));
   }
 
-  RunEvent buildRun(
-      ParentRunFacet applicationParentRunFacet,
-      RunEventBuilder runEventBuilder,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      SparkListenerStageSubmitted event) {
-    Stage stage = stageMap.get(event.stageInfo().stageId());
-    RDD<?> rdd = stage.rdd();
-
-    List<Object> nodes = new ArrayList<>();
-    nodes.addAll(Arrays.asList(event.stageInfo(), stage));
-
-    nodes.addAll(Rdds.flattenRDDs(rdd));
-
-    return populateRun(
-        applicationParentRunFacet,
-        runEventBuilder,
-        openLineageContext.getRunUuid(),
-        jobBuilder,
-        jobFacetsBuilder,
-        nodes);
-  }
-
-  RunEvent buildRun(
-      ParentRunFacet applicationParentRunFacet,
-      RunEventBuilder runEventBuilder,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      SparkListenerStageCompleted event) {
-    Stage stage = stageMap.get(event.stageInfo().stageId());
-    RDD<?> rdd = stage.rdd();
-
-    List<Object> nodes = new ArrayList<>();
-    nodes.addAll(Arrays.asList(event.stageInfo(), stage));
-
-    nodes.addAll(Rdds.flattenRDDs(rdd));
-
-    return populateRun(
-        applicationParentRunFacet,
-        runEventBuilder,
-        openLineageContext.getRunUuid(),
-        jobBuilder,
-        jobFacetsBuilder,
-        nodes);
-  }
-
-  RunEvent buildRun(
-      ParentRunFacet applicationParentRunFacet,
-      RunEventBuilder runEventBuilder,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      SparkListenerSQLExecutionStart event) {
-    return buildRun(
-        applicationParentRunFacet,
-        runEventBuilder,
-        openLineageContext.getRunUuid(),
-        jobBuilder,
-        jobFacetsBuilder,
-        event,
-        Optional.empty());
-  }
-
-  RunEvent buildRun(
-      ParentRunFacet applicationParentRunFacet,
-      RunEventBuilder runEventBuilder,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      SparkListenerSQLExecutionEnd event) {
-    return buildRun(
-        applicationParentRunFacet,
-        runEventBuilder,
-        openLineageContext.getRunUuid(),
-        jobBuilder,
-        jobFacetsBuilder,
-        event,
-        Optional.empty());
-  }
-
-  RunEvent buildRun(
-      ParentRunFacet parentRunFacet,
-      RunEventBuilder runEventBuilder,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      SparkListenerJobStart event) {
-    return buildRun(
-        parentRunFacet,
-        runEventBuilder,
-        openLineageContext.getRunUuid(),
-        jobBuilder,
-        jobFacetsBuilder,
-        event,
-        Optional.ofNullable(jobMap.get(event.jobId())));
-  }
-
-  RunEvent buildRun(
-      ParentRunFacet applicationParentRunFacet,
-      RunEventBuilder runEventBuilder,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      SparkListenerJobEnd event) {
-    return buildRun(
-        applicationParentRunFacet,
-        runEventBuilder,
-        openLineageContext.getRunUuid(),
-        jobBuilder,
-        jobFacetsBuilder,
-        event,
-        Optional.ofNullable(jobMap.get(event.jobId())));
-  }
-
-  RunEvent buildRun(
-      ParentRunFacet parentRunFacet,
-      RunEventBuilder runEventBuilder,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      SparkListenerApplicationStart event) {
-    return buildRun(
-        parentRunFacet,
-        runEventBuilder,
-        openLineageContext.getApplicationUuid(),
-        jobBuilder,
-        jobFacetsBuilder,
-        event,
-        Optional.empty());
-  }
-
-  RunEvent buildRun(
-      ParentRunFacet applicationParentRunFacet,
-      RunEventBuilder runEventBuilder,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      SparkListenerApplicationEnd event) {
-    return buildRun(
-        applicationParentRunFacet,
-        runEventBuilder,
-        openLineageContext.getApplicationUuid(),
-        jobBuilder,
-        jobFacetsBuilder,
-        event,
-        Optional.empty());
-  }
-
-  private RunEvent buildRun(
-      ParentRunFacet applicationParentRunFacet,
-      RunEventBuilder runEventBuilder,
-      UUID runId,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      Object event,
-      Optional<ActiveJob> job) {
-    List<Object> nodes = new ArrayList<>();
-    nodes.add(event);
-    job.ifPresent(
-        j -> {
-          nodes.add(j);
-          nodes.addAll(Rdds.flattenRDDs(j.finalStage().rdd()));
-        });
-
-    return populateRun(
-        applicationParentRunFacet, runEventBuilder, runId, jobBuilder, jobFacetsBuilder, nodes);
-  }
-
-  private RunEvent populateRun(
-      ParentRunFacet applicationParentRunFacet,
-      RunEventBuilder runEventBuilder,
-      UUID runId,
-      JobBuilder jobBuilder,
-      OpenLineage.JobFacetsBuilder jobFacetsBuilder,
-      List<Object> nodes) {
+  RunEvent buildRun(OpenLineageRunEventContext context) {
     OpenLineage openLineage = openLineageContext.getOpenLineage();
-
+    List<Object> nodes = context.loadNodes(stageMap, jobMap);
+    UUID runId = context.getOverwriteRunId().orElse(openLineageContext.getRunUuid());
     RunFacetsBuilder runFacetsBuilder = openLineage.newRunFacetsBuilder();
 
-    runFacetsBuilder.parent(applicationParentRunFacet);
-    OpenLineage.JobFacets jobFacets = buildJobFacets(nodes, jobFacetBuilders, jobFacetsBuilder);
+    runFacetsBuilder.parent(context.getApplicationParentRunFacet());
+    OpenLineage.JobFacets jobFacets =
+        buildJobFacets(nodes, jobFacetBuilders, context.getJobFacetsBuilder());
     List<InputDataset> inputDatasets = buildInputDatasets(nodes);
     List<OutputDataset> outputDatasets = buildOutputDatasets(nodes);
     openLineageContext
@@ -385,14 +208,15 @@ class OpenLineageRunEventBuilder {
 
     RunFacets runFacets = buildRunFacets(nodes, runFacetBuilders, runFacetsBuilder);
     OpenLineage.RunBuilder runBuilder = openLineage.newRunBuilder().runId(runId).facets(runFacets);
-    runEventBuilder
+    context
+        .getRunEventBuilder()
         .run(runBuilder.build())
-        .job(jobBuilder.facets(jobFacets).build())
+        .job(context.getJobBuilder().facets(jobFacets).build())
         .inputs(RemovePathPatternUtils.removeInputsPathPattern(openLineageContext, inputDatasets))
         .outputs(
             RemovePathPatternUtils.removeOutputsPathPattern(openLineageContext, outputDatasets));
 
-    return runEventBuilder.build();
+    return context.getRunEventBuilder().build();
   }
 
   private List<InputDataset> buildInputDatasets(List<Object> nodes) {
@@ -405,10 +229,12 @@ class OpenLineageRunEventBuilder {
                 log.debug("Physical plan executed {}", qe.executedPlan().toJSON());
               }
             });
-    log.debug(
-        "Visiting query plan {} with input dataset builders {}",
-        openLineageContext.getQueryExecution(),
-        inputDatasetBuilders);
+    if (log.isDebugEnabled()) {
+      log.debug(
+          "Visiting query plan {} with input dataset builders {}",
+          openLineageContext.getQueryExecution(),
+          inputDatasetBuilders);
+    }
 
     Function1<LogicalPlan, Collection<InputDataset>> inputVisitor =
         visitLogicalPlan(PlanUtils.merge(inputDatasetQueryPlanVisitors));
@@ -482,10 +308,12 @@ class OpenLineageRunEventBuilder {
   }
 
   private List<OutputDataset> buildOutputDatasets(List<Object> nodes) {
-    log.debug(
-        "Visiting query plan {} with output dataset builders {}",
-        openLineageContext.getQueryExecution(),
-        outputDatasetBuilders);
+    if (log.isDebugEnabled()) {
+      log.debug(
+          "Visiting query plan {} with output dataset builders {}",
+          openLineageContext.getQueryExecution(),
+          outputDatasetBuilders);
+    }
     Function1<LogicalPlan, Collection<OutputDataset>> visitor =
         visitLogicalPlan(
             PlanUtils.merge(